requirements.txt
# Azure AI Services
azure-ai-formrecognizer==3.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.34.0
azure-search-documents==11.4.0
azure-ai-ml==1.12.0

# Azure OpenAI
openai==1.6.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0
pypdf2==3.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0